Move to module pattern
Instead of requiring a Simctl object, we now use a module pattern instead. For
example, instead of listing devices by doing this:

    simctl = Simctl()
    print simctl.devices()

We can now run:

    print xcrun.simctl.listall.devices()

These commands are closer to the actual `xcrun` command and have the advantage
that we don't need to preload everything when we create the `Simctl` object.